+++ app/controllers/dorsale/customer_vault/links_controller.rb
@@ -8,20 +8,20 @@ class Dorsale::CustomerVault::LinksController < ::Dorsale::CustomerVault::Applic
   ]
 
   def new
-    authorize! :update, @person
+    authorize @person, :update?
 
     @link   ||= model.new
     @people ||= current_user_scope.people
   end
 
   def edit
-    authorize! :update, @person
+    authorize @person, :update?
 
     @link = model.find(params[:id])
   end
 
   def create
-    authorize! :update, @person
+    authorize @person, :update?
 
     params = link_params
     bob = params[:bob].split("-")
@@ -37,7 +37,7 @@ class Dorsale::CustomerVault::LinksController < ::Dorsale::CustomerVault::Applic
   end
 
   def update
-    authorize! :update, @person
+    authorize @person, :update?
 
     @link = model.find(params[:id])
 
@@ -50,14 +50,14 @@ class Dorsale::CustomerVault::LinksController < ::Dorsale::CustomerVault::Applic
   end
 
   def destroy
-    authorize! :update, @person
+    authorize @person, :update?
 
     @link = model.find(params[:id])
 
     if @link.destroy
-      flash[:notice] = t("messages.links.destroy_ok")
+      flash[:notice] = t("messages.links.delete_ok")
     else
-      flash[:alert] = t("messages.links.destroy_error")
+      flash[:alert] = t("messages.links.delete_error")
     end
 
     redirect_to back_url
